+++ src/cca/scripts/background.ts
@@ -1,5 +1,4 @@
 /// <reference path='../../../third_party/cordova/splashscreen.d.ts'/>
-/// <reference path='../../../third_party/cordova/webintents.d.ts'/>		
 /// <reference types="chrome/chrome-app" />
 
 /**
@@ -9,8 +8,6 @@
  * loaded.
  */
 
-import * as uproxy_core_api from '../../interfaces/uproxy_core_api';
-
 import { ServerListPage } from '../ui_components/server_list';
 import { UproxyServerRepository } from './uproxy_server';
 import { MakeCoreConnector } from './cordova_core_connector';
@@ -18,30 +15,13 @@ import { GetGlobalTun2SocksVpnDevice } from './tun2socks_vpn_device';
 import * as vpn_device from '../model/vpn_device';
 import * as intents from './intents';
 
-declare const freedom: freedom.FreedomInCoreEnv;
-
-export interface OnEmitModule extends freedom.OnAndEmit<any,any> {};
-export interface OnEmitModuleFactory extends
-  freedom.FreedomModuleFactoryManager<OnEmitModule> {};
-
-console.log('Loading core');
-export var uProxyAppChannel = freedom(
-    'generic_core/freedom-module.json',
-    <freedom.FreedomInCoreEnvOptions>{
-      'logger': 'lib/loggingprovider/freedom-module.json',
-      'debug': 'debug',
-      'portType': 'worker'
-    }
-).then((uProxyModuleFactory:OnEmitModuleFactory) => {
-  console.log('Core loading complete');
-  return uProxyModuleFactory();
-});
-
 // We save this reference to allow inspection of the context state from the browser debuggin tools.
 (window as any).context = this;
 
+// TODO: Get rid of core connector and talk to the core directly.
 let core = MakeCoreConnector();
 
+// Create UproxyServerRepository.
 let serversPromise = GetGlobalTun2SocksVpnDevice().then((vpnDevice) => {
   console.debug('Device supports VPN');
   return vpnDevice;
@@ -52,6 +32,7 @@ let serversPromise = GetGlobalTun2SocksVpnDevice().then((vpnDevice) => {
   return new UproxyServerRepository(core, vpnDevice);
 });
 
+// Create UI.
 let serversListPagePromise: Promise<ServerListPage> = new Promise((resolve, reject) => {
   chrome.app.runtime.onLaunched.addListener(() => {
     console.debug('Chrome onLaunched fired');
@@ -66,6 +47,7 @@ let serversListPagePromise: Promise<ServerListPage> = new Promise((resolve, reje
   });
 });
 
+// Register for url intents.
 Promise.all([serversListPagePromise, intents.GetGlobalIntentInterceptor()])
 .then(([serverListPage, intentInterceptor]) => {
   intentInterceptor.addIntentListener((url: string) => {
